Make autoconversion from sbinary -> jsonnew only apply in sbt 1.x
