+++ src/main/scala-sbt-0.13/net/virtualvoid/sbt/graph/ModuleGraphProtocolCompat.scala
@@ -1,0 +1,3 @@
+package net.virtualvoid.sbt.graph
+
+trait ModuleGraphProtocolCompat
